pyproject.toml
[project]
name = "comfyui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "customtkinter",
    "websocket-client",
    "pillow",
]
